local_requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
matplotlib>=3.8.0
reportlab>=4.0.0
camelot-py>=0.11.0
ghostscript>=0.7
pytesseract>=0.3.10
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
seaborn>=0.13.0
statsmodels>=0.14.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
langchain>=0.0.350
pycairo>=1.25.0